requirements.txt
elasticsearch==8.11.1
requests==2.31.0
python-dateutil==2.8.2


